auth-jwt/api/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
ecdsa==0.19.1
Flask==3.1.0
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pyasn1==0.4.8
python-jose==3.4.0
rsa==4.9
six==1.17.0
Werkzeug==3.1.3
